# openapi.json
{"openapi":"3.1.0","info":{"title":"Сервис по расчёту стоимости страхования","description":"Расчёт стоимости страхования в зависимости от типа грузаи объявленной стоимости (ОС)","version":"0.1.0"},"paths":{"/cargo_rate/all_tariff":{"get":{"tags":["Cargo Rate"],"summary":"Get Current Tariff All Date","operationId":"get_current_tariff_all_date_cargo_rate_all_tariff_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TariffFile"}}}}}}},"/cargo_rate/upload_tariff_file":{"post":{"tags":["Cargo Rate"],"summary":"Upload Tariff File","operationId":"upload_tariff_file_cargo_rate_upload_tariff_file_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_tariff_file_cargo_rate_upload_tariff_file_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cargo_rate/upload_tariff_data":{"post":{"tags":["Cargo Rate"],"summary":"Upload Tariff Date","operationId":"upload_tariff_date_cargo_rate_upload_tariff_data_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TariffFile"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cargo_rate/get_cost_of_insurance":{"post":{"tags":["Cargo Rate"],"summary":"Get Cost Of Insurance","operationId":"get_cost_of_insurance_cargo_rate_get_cost_of_insurance_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CostInsurance"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_tariff_file_cargo_rate_upload_tariff_file_post":{"properties":{"tariff_file":{"type":"string","format":"binary","title":"Tariff File"}},"type":"object","required":["tariff_file"],"title":"Body_upload_tariff_file_cargo_rate_upload_tariff_file_post"},"CostInsurance":{"properties":{"date":{"type":"string","title":"Date","example":"2020-06-01"},"cargo_type":{"type":"string","title":"Cargo Type","example":"Glass"},"declared_value":{"type":"number","title":"Declared Value","example":10000}},"type":"object","required":["date","cargo_type","declared_value"],"title":"CostInsurance"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RateItem":{"properties":{"cargo_type":{"type":"string","title":"Cargo Type"},"rate":{"type":"string","title":"Rate"}},"type":"object","required":["cargo_type","rate"],"title":"RateItem"},"TariffFile":{"additionalProperties":{"items":{"$ref":"#/components/schemas/RateItem"},"type":"array"},"type":"object","title":"TariffFile","example":{"2020-06-01":[{"cargo_type":"Glass","rate":"0.04"},{"cargo_type":"Other","rate":"0.01"}],"2020-07-01":[{"cargo_type":"Glass","rate":"0.035"},{"cargo_type":"Other","rate":"0.015"}]}},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}